refactor: separate runBenchmark into dedicated module
Extract runBenchmark and related utilities from index.mts into run-benchmark.mts
to improve code organization and module dependencies. This refactoring:

- Creates a dedicated module for benchmark execution (run-benchmark.mts)
- Simplifies index.mts to only export public API
- Fixes import paths and module references
- Improves separation of concerns between modules
